Store environment to flash on first boot
Store the env to flash if the env in flash doesn't have the key 'fromflash'
defined, this should only happen on the first boot after the bootloader
is installed.
This simplifies the bootstraping in the factory and modifications from
linux without using the serial interface nor having to know from linux which
variables should be in the environment.